feature(plugin): add OpenAPI Document paths filtering
