--- packages/plugin/src/lib/open-api/getDocumentServices.ts
@@ -1,14 +1,0 @@
-import { getServices, ServiceOutputOptions } from './getServices.js';
-import { OpenAPISchemaType } from './OpenAPISchemaType.js';
-
-export const getDocumentServices = async ({
-  schema,
-  servicesGlob,
-  output,
-}: {
-  schema: OpenAPISchemaType;
-  servicesGlob: string[] | undefined;
-  output: ServiceOutputOptions;
-}) => {
-  return getServices(schema, output, servicesGlob);
-};
